Add assertions and bug fix
Add some assert for the commands. Also fixed the bugs due to difference
in compatability
