--- src/main/java/java/tasks/ToDoTask.java
@@ -1,22 +1,0 @@
-package tasks;
-
-public class ToDoTask extends Task {
-    public ToDoTask(String description) {
-        super(description);
-    }
-
-    @Override
-    protected String getTypeIcon() {
-        return "[T]";
-    }
-
-    @Override
-    public String toString() {
-        return getTypeIcon() + getStatusIcon() + " " + description;
-    }
-
-    @Override
-    public String toStorageString() {
-        return "T | " + (isDone ? "1" : "0") + " | " + description;
-    };
-}
