--- src/main/java/java/ui/DarrenAssistantException.java
@@ -1,7 +1,0 @@
-package ui;
-
-public class DarrenAssistantException extends Exception{
-    public DarrenAssistantException(String message) {
-        super(message);
-    }
-}
